8303950: use new paintContainer(Graphics) method
This is an alternative proposal for resolving 8303950 without modifying Window.java . Here we offer a new hook in Container (the new `protected void paintContainer(Graphics)` method for subclasses to directly access.

This is my less-favorite approach, because it's based on the assumption that we *know* that JWindow, JFrame, and JDialog want to bypass their `super.paint(Graphics)` logic. If someday Window.java is revised and new logic is added that we want/need: we'll miss out on it. Because this PR's "fix" is based on taking advantage of current implementation details more than theoretical design.