8303950: use new Window#paint(Graphics, boolean) method
This is a second alternative proposal for resolving 8303950.

Pros: the method signatures better signal the real intention/implementation here.

Cons: this still adds a protected method to Window.java, which I'd like to avoid. This new method isn't really going to be leveraged by any existing subclasses except the ones in this PR.